+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.7.1
 - Fix table component alignment and styling
 - Fix toggle button styling
+- Miscellanous code improvements
 
 # v0.7.1
 - Add shading to table component