+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improve container index table
 - Fix bug with ammo not updating after deleting shot group
 - Replace ammo "added on" with "purchased on"
+- Miscellaneous wording improvements
 
 # v0.7.1
 - Add shading to table component