+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Miscellaneous wording improvements
 - Update translations
 
-# v0.7.1
+# v0.7.0
 - Add shading to table component
 - Fix chart to sum by day
 - Fix whitespace when copying invite url